refactor: cache zasm script size

+++ src/zq/zq_class.cpp
@@ -12940,7 +12940,7 @@ int32_t write_one_ffscript(PACKFILE *f, zquestheader *Header, int32_t i, script_
     Header=Header;
     i=i;
     
-    int32_t num_commands = (*script)->size();
+    size_t num_commands = (*script)->size;
     
     if(!p_iputl(num_commands,f))
     {
